extract hazel messages to subfolder
